Separate-out components.
